+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Automatic expansion
 - Replace incompatible types
+- Array insertion using index number of `-1`
 
 ## [1.0.0] - 2018-03-21
 ### Added